requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0
supabase>=1.0.0
psycopg2-binary>=2.9
httpx>=0.24.0
